src/main/java/com/database/Spring_PostGres/dao/impl/AuthorDaoImpl.java
package com.database.Spring_PostGres.dao.impl;

import com.database.Spring_PostGres.dao.AuthorDao;
import com.database.Spring_PostGres.domain.Author;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.jdbc.core.RowMapper;
import org.springframework.stereotype.Component;

import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.List;
import java.util.Optional;

@Component
public class AuthorDaoImpl implements AuthorDao {
    private final JdbcTemplate jdbcTemplate;

    public AuthorDaoImpl(final JdbcTemplate jdbcTemplate) {
        this.jdbcTemplate = jdbcTemplate;
    }

    @Override
    public void create(Author author) {
        jdbcTemplate.update(
                "INSERT INTO authors (id,age,name) VALUES(?,?,?)",
                author.getId(),author.getAge(),author.getName()
        );
    }

    @Override
    public Optional<Author> findOne(long authorId) {
        List<Author> results = jdbcTemplate.query("SELECT id,age,name FROM author WHERE id = ? LIMIT 1", new AuthorRowMapper(), authorId);
        return results.stream().findFirst();
    }

    public static class AuthorRowMapper implements RowMapper<Author> {
        @Override
        public Author mapRow(ResultSet rs, int rowNum) throws SQLException {
            return Author.builder()
                    .id(rs.getLong("id"))
                    .age(rs.getInt("age"))
                    .name(rs.getString("name"))
                    .build();
        }
    }
}
